pyproject.toml
[project]
name = "commerce-ai-orchestrator"
version = "0.1.0"
description = "Multi-agent system with CSV and SQL analysis capabilities using MCP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # From agents project
    "semantic-kernel",
    "python-a2a",
    "fastapi",
    "uvicorn",
    "a2a-sdk[sqlite]",
    "mcp[cli]>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.17.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "openpyxl>=3.1.0",
    "xlrd>=2.0.0",
    "python-dateutil>=2.8.0",
    "pytz>=2023.3",
    "google-auth>=2.16.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.100.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "asyncio>=3.4.3",
    "httpx>=0.25.0",
    "pydantic-ai>=0.0.15",
    "rich>=13.0.0",
    "mysql-connector-python",
    "fastmcp>=0.1.0",
]

[project.scripts]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["a2a_azure"] 
